0.965 ナビゲーションの点検と修正、仕様書を追加
ナビゲーション周りを通しで点検し、見つかった不具合を直した。

■ 判断が地図画面を出している間しか動かなかった
Auto10km の折返し判定・コース逸脱警報・旋回角速度の更新が MODE_MAP の
分岐の中にあり、設定画面や各詳細画面を開いている間は全部止まっていた。
とくに INTO→AWAY（目的地から 1.5km 以内）は「そのとき近くにいるか」で
判定するので、地図画面を離れている間に通り過ぎると二度と成立しない。
nav_alarm_tick() に切り出して loop() の最上位へ移した
（電池警告を移したのと同じ理由。CLAUDE.md 参照）。
あわせて、設定画面はボタン操作が 3 分無ければ自動で地図画面へ戻す。
設定画面を出したまま離陸してしまった場合の保険で、値変更モードのまま
時間切れになっても CallbackExit（Auto10km と目的地の組み合わせチェック）は通す。

■ コース警報が 1 周期古いズレ角で判定していた
update_course_warning() が steer_angle を読んだ後に steer_angle を計算し、
その元になる truec はさらに後の nav_update() で更新していた。
nav_update() → steer_angle → update_course_warning() の順に直し、
入れ替え禁止の理由をコメントに残した。

■ deg2rad が double 版を持ちながら float 版に落ちていた
double 版がヘッダに宣言されておらず、calculateDistanceKm() や nav_update() は
引数が double でも float 版で計算していた（引数型 double が実質的に無意味だった）。
float 版を削除して double 1 本に統一。PI が double リテラルなので掛け算は
元から double で行われており、速度の代償は無い（Flash はむしろ減った）。

■ 目的地配列の解放が飛行中に Core0 と競合し得た
起動時に SD を掴めず飛行中に認識した場合だけ、Core1 の init_destinations() が
delete[] と .cords = nullptr を書く一方で Core0 が同じ配列を読んでいた。
解放をやめてポインタの差し替えだけにし、destinations_count を最後に公開する
順序にした。捨てるのは起動あたり最大 1 世代・約 100 バイトで、
destinations_loaded が二度目以降の呼び出しを止めるため増えない。

■ 測位前に意味のない TC が出ていた
内部位置が (0,0) のままなので 3 桁の方位がそれらしく読めてしまっていた。
距離の ---km と同じ条件で TC--- にした。

docs/pons_navigation.md を新規作成（方位と距離の出し方、3 つのナビモード、
Auto10km の状態機械、コース警報の積算式、琵琶湖コースの描画）。
README と CLAUDE.md から辿れるようにした。

RELEASE と非 RELEASE の両構成でビルド確認済み。新規の警告なし。

+++ navdata.h
@@ -6,7 +6,7 @@
 //           内蔵ポリゴン地図・目的地・飛行コースモードの定義、
 //           各飛行地点（琵琶湖・白浜・笠岡・富士川・東京湾など）の座標。
 // Author  : MasaoC (@masao_mobile)
-// Updated : 2026/09/10
+// Updated : 2026/09/13
 // ============================================================
 #include "settings.h"
 
@@ -105,7 +105,7 @@
   // libc の strdup() と混ざらないよう、名前を分けてここで宣言しておく。
   char* pons_strdup(const char* str);
 
-  float deg2rad(float degrees);
+  double deg2rad(double degrees);
   double rad2deg(double rad);
   extern int truec;
